update to new icu
git-svn-id: https://unicode.org/repos/unicodetools/trunk@281 13e8329f-0b23-4da4-9fe8-d0f6fe080806

+++ org/unicode/jsp/BidiCharMap.java
@@ -63,7 +63,7 @@ class BidiCharMap {
     asciiHackMap.put('"',NSM);
     asciiHackMap.put('_',BN);
   }
-  
+
   boolean asciiHack;
 
   public BidiCharMap (boolean asciiHack) {
@@ -78,12 +78,12 @@ class BidiCharMap {
     if (asciiHack2) {
       Byte result = (Byte) asciiHackMap.getValue(codepoint);
       if (result != null) {
-        return (byte) result;
+        return result;
       }
     }
     return mapIcuToRefNum[UCharacter.getIntPropertyValue(codepoint, UProperty.BIDI_CLASS)];
   }
-  
+
   public byte getBidiClass(int codepoint) {
     return getBidiClass(codepoint, asciiHack);
   }
